requirement.txt
opencv-python
SpeechRecognition
pyttsx3
pyaudio
langchain
langchain-google-genai
tensorflow
pillow
python-dotenv
pyserial
